feat: rework as a transport
