+++ implementation-notes.md
@@ -124,5 +124,4 @@ In order to minimize nesting, the first example is preferred
                     -  A generic circuit address will be added to the peers multiaddr list
                         - i.e. `/p2p-circuit/ipfs/QmDest`
             - If another transport is available, then use that instead of the relay
-
