+++ src/index.js
@@ -1,5 +1,5 @@
-'use strict'
-
-exports = module.exports = require('./relay')
-exports.Circuit = require('./transport')
-exports.Listener = require('./transport/listener')
+module.exports = {
+  Circuit: require('./circuit'),
+  Dialer: require('./dialer'),
+  multicodec: require('./multicodec')
+}
