+++ src/multicodec.js
@@ -1,0 +1,1 @@
+module.exports = '/ipfs/relay/circuit/1.0.0'
